requirements.txt
PyYAML>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.0.0
anthropic>=0.18.0
lxml>=4.9.0
